% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRO_interface.R
\name{assemble_protein_complex}
\alias{assemble_protein_complex}
\title{assemble_protein_complex}
\usage{
assemble_protein_complex(organism_id)
}
\arguments{
\item{organism_id}{}
}
\value{

}
\description{
assemble_protein_complex
}
