% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_queries.R
\name{build_allele_group}
\alias{build_allele_group}
\title{build_allele_group}
\usage{
build_allele_group(allele_selection)
}
\arguments{
\item{allele_selection}{}
}
\value{

}
\description{
build_allele_group
}
