% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{check_geographics}
\alias{check_geographics}
\title{check_geographics, stops if input not adequate}
\usage{
check_geographics(country, region, ethnic)
}
\arguments{
\item{country}{country used for allele frequency selection}

\item{region}{geographical region used for allele frequency selection}

\item{ethnic}{ethical origin used for allele frequency selection}
}
\value{
TRUE
}
\description{
check_geographics, stops if input not adequate
}
\keyword{internal}
