% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{check_hla_locus}
\alias{check_hla_locus}
\title{check_hla_locus, stops if input not adequate}
\usage{
check_hla_locus(hla_locus)
}
\arguments{
\item{hla_locus}{locus of hla frequencies to query}
}
\value{
TRUE
}
\description{
check_hla_locus, stops if input not adequate
}
\keyword{internal}
