% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{check_hla_selection}
\alias{check_hla_selection}
\title{check_hla_selection, stops if input not adequate}
\usage{
check_hla_selection(hla_selection, query_type)
}
\arguments{
\item{hla_selection}{HLA alleles used for selection}

\item{query_type}{["allele"|"haplotype"] type of AFND query}
}
\value{
TRUE
}
\description{
check_hla_selection, stops if input not adequate
}
