% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{check_population}
\alias{check_population}
\title{check_population, stops if input not adequate}
\usage{
check_population(hla_population)
}
\arguments{
\item{hla_population}{population id}
}
\value{
TRUE
}
\description{
check_population, stops if input not adequate
}
