% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{check_sample_size}
\alias{check_sample_size}
\title{check_sample_size, stops if input not adequate}
\usage{
check_sample_size(hla_sample_size_pattern, hla_sample_size)
}
\arguments{
\item{hla_sample_size_pattern}{one of "bigger_than", "equal", "less_than", "less_equal_than", "bigger_equal_than","different"}

\item{hla_sample_size}{integer number used for population size}
}
\value{
TRUE
}
\description{
check_sample_size, stops if input not adequate
}
