% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{check_standard}
\alias{check_standard}
\title{check_standard, stops if input not adequate}
\usage{
check_standard(standard)
}
\arguments{
\item{standard}{one of "a" - all,"s" - silver,"g" - gold}
}
\value{
TRUE
}
\description{
check_standard, stops if input not adequate
}
