% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACUI_interface.R
\name{create_encode_handle}
\alias{create_encode_handle}
\title{create_encode_handle}
\usage{
create_encode_handle(allele_names)
}
\arguments{
\item{allele_names}{}
}
\value{
curl handle
}
\description{
create_encode_handle
}
