% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACUI_interface.R
\name{encode_MAC}
\alias{encode_MAC}
\title{encode_MAC}
\usage{
encode_MAC(allele_list)
}
\arguments{
\item{allele_list}{}
}
\value{
encoded MAC
}
\description{
encode_MAC
}
\examples{
allele_list <- c("A*01:01:01", "A*02:01:01", "A*03:01")
encode_MAC(allele_list)

}
