% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{extract_population_id}
\alias{extract_population_id}
\title{extract_population_id}
\usage{
extract_population_id(data)
}
\arguments{
\item{data}{html from AFND website}
}
\value{
population ids
}
\description{
extract the population ids from the html result
}
\keyword{internal}
