% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{extract_population_info}
\alias{extract_population_info}
\title{extract_population_info}
\usage{
extract_population_info(data)
}
\arguments{
\item{data}{html input page}
}
\value{
population information
}
\description{
extract_population_info
}
