% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{extract_population_name}
\alias{extract_population_name}
\title{extract_population_name}
\usage{
extract_population_name(data)
}
\arguments{
\item{data}{html input page}
}
\value{
population name
}
\description{
extract_population_name
}
