% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{extract_sample_info}
\alias{extract_sample_info}
\title{extract_sample_info}
\usage{
extract_sample_info(data)
}
\arguments{
\item{data}{html input page}
}
\value{
sample information
}
\description{
extract_sample_info
}
