% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACUI_interface.R
\name{fetch_encoded_MAC}
\alias{fetch_encoded_MAC}
\title{fetch_encoded_MAC}
\usage{
fetch_encoded_MAC(handle)
}
\arguments{
\item{handle}{}
}
\value{
curl handle fetch
}
\description{
fetch_encoded_MAC
}
