% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources_input.R
\name{get_external_file}
\alias{get_external_file}
\title{get_external_file}
\usage{
get_external_file(file, skip = 0, delim = "\\t", col_names = TRUE)
}
\arguments{
\item{file}{Indicated the file that will be read}

\item{skip}{integer indicating how many lines to skip when reading URL}

\item{delim}{pattern used for delim 
(passed to \code{delim} of read functions)}

\item{col_names}{list of colnames to use}
}
\value{
returns a the content of the file. The format of the return object
depends on the read_method that was used.
}
\description{
get_external_file
}
\keyword{internal}
