% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature_queries.R
\name{get_mhcpan_input}
\alias{get_mhcpan_input}
\title{get_mhcpan_input}
\usage{
get_mhcpan_input(name_list, mhc_class = c("MHC-I", "MHC-II"))
}
\arguments{
\item{name_list}{}
}
\value{
Protein chain list as formatted for MHCpan input
}
\description{
get_mhcpan_input
}
