% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{get_nb_pages}
\alias{get_nb_pages}
\title{get_nb_pages}
\usage{
get_nb_pages(page_tbl)
}
\arguments{
\item{page_tbl}{html input page}
}
\value{
integer number of pages in the html input
}
\description{
get_nb_pages
}
