% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature_queries.R
\name{get_serotypes}
\alias{get_serotypes}
\title{get_serotypes}
\usage{
get_serotypes(allele_list, organism = "human", mhc_type = c("MHC-I", "MHC-II"))
}
\arguments{
\item{allele_list}{}
}
\value{
serotype
}
\description{
get_serotypes
}
\examples{
allele_list <- c("A*01:01:01","B*27:01")
get_serotypes(allele_list, mhc_type = "MHC-I")

}
