% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{get_valid_geographics}
\alias{get_valid_geographics}
\title{get_valid_geographics}
\usage{
get_valid_geographics()
}
\value{
list of valid countries, regions and ethnic origin
}
\description{
get a list of valid countries, regions and ethnic origins
}
\keyword{internal}
