% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRO_interface_helper.R
\name{get_valid_organisms}
\alias{get_valid_organisms}
\title{get_valid_organisms
query the possible species}
\usage{
get_valid_organisms()
}
\value{
a list of organisms that are part of the annotation functions
}
\description{
get_valid_organisms
query the possible species
}
