% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature_queries.R
\docType{data}
\name{human_protein_complex_table}
\alias{human_protein_complex_table}
\title{mro.obo}
\format{
An object of class \code{data.frame} with 12386 rows and 8 columns.
}
\usage{
human_protein_complex_table
}
\description{
mro.obo
}
\details{
\code{human_protein_complex_table}: human_protein_complex_table.
}
\keyword{datasets}
