% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_resources_input.R
\docType{data}
\name{mro.obo}
\alias{mro.obo}
\title{MRO}
\format{
An object of class \code{ontology_index} of length 18.
}
\usage{
mro.obo
}
\description{
\code{mro.obo}: MHC Restriction Ontology obo file.
}
\keyword{datasets}
