% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{parse_allele_freq_html}
\alias{parse_allele_freq_html}
\title{parse_allele_freq_html}
\usage{
parse_allele_freq_html(allele_freq_table)
}
\arguments{
\item{allele_freq_table}{allele frequency table parsed from AFND website}
}
\value{
allele_freq_table reformatted
}
\description{
format the allele frequency table and select columns of interest
}
