% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{parse_haplotype_freq_html}
\alias{parse_haplotype_freq_html}
\title{parse_haplotype_freq_html}
\usage{
parse_haplotype_freq_html(haplotype_freq_table)
}
\arguments{
\item{haplotype_freq_table}{haplotype frequency table parsed
from AFND website}
}
\value{
haplotype_freq_table reformatted
}
\description{
format the haplotype frequency table and select 
columns of interest
}
