% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_allele_frequency}
\alias{plot_allele_frequency}
\title{plot_allele_frequency}
\usage{
plot_allele_frequency(allele_frequency)
}
\arguments{
\item{allele_frequency}{}
}
\value{
ggplot
}
\description{
plot_allele_frequency
}
