% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_queries.R
\name{query_population_detail}
\alias{query_population_detail}
\title{Query population metainformation}
\usage{
query_population_detail(population_ids)
}
\arguments{
\item{population_ids}{List of numeric identifiers of the population that will be used for filtering. Thie identifier is defined by the Allele Frequency Net Database.}
}
\value{
data.frame object containing the result of the population detail query
}
\description{
Query population metainformation
}
\examples{
library(AFNDquery)
population_detail <- query_population_detail(0001986)
}
