% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{read_complete_freq_table}
\alias{read_complete_freq_table}
\title{read_complete_freq_table}
\usage{
read_complete_freq_table(url, type)
}
\arguments{
\item{url}{URL of the website containing frequency table}

\item{type}{["allele"|"haplotype"]}
}
\value{
frequency table
}
\description{
read_complete_freq_table
}
