% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{read_population_detail}
\alias{read_population_detail}
\title{read_population_detail}
\usage{
read_population_detail(url, population_id)
}
\arguments{
\item{url}{url of the page containing population information}

\item{population_id}{numeric population identifier}
}
\value{
population information
}
\description{
read_population_detail
}
\keyword{internal}
