% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRO_interface.R
\name{retrieve_chain_lookup_table}
\alias{retrieve_chain_lookup_table}
\title{retrieve_chain_lookup_table}
\usage{
retrieve_chain_lookup_table(organism)
}
\arguments{
\item{organism}{}
}
\value{

}
\description{
retrieve_chain_lookup_table
}
\examples{
retrieve_chain_lookup_table("mouse")

}
