% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{verify_parameters}
\alias{verify_parameters}
\title{verify_parameters}
\usage{
verify_parameters(
  hla_locus,
  hla_selection,
  hla_population,
  hla_country,
  hla_region,
  hla_ethnic,
  hla_sample_size_pattern,
  hla_sample_size,
  standard = "a",
  query_type
)
}
\arguments{
\item{hla_locus}{locus of hla frequencies to query}

\item{hla_selection}{HLA alleles used for selection}

\item{hla_population}{population id}

\item{hla_country}{country used for allele frequency selection}

\item{hla_region}{geographical region used for allele frequency selection}

\item{hla_ethnic}{ethical origin used for allele frequency selection}

\item{hla_sample_size_pattern}{one of "bigger_than", "equal", "less_than", 
"less_equal_than", "bigger_equal_than","different"}

\item{hla_sample_size}{integer number used for population size}

\item{standard}{one of "a" - all,"s" - silver,"g" - gold}

\item{query_type}{"allele" or "haplotype"}
}
\value{
boolean to indicate, whether tests were passed
}
\description{
verify_parameters
}
