% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRO_interface.R
\name{assemble_protein_complex}
\alias{assemble_protein_complex}
\title{Assemble protein complex}
\usage{
assemble_protein_complex(organism)
}
\arguments{
\item{organism}{Organism for which the lookup should be built (e.g. 
"human", "mouse", ...). The list of valid 
organisms can be found using the function \code{get_valid_organisms}}
}
\value{
a data frame with the MHC complexes annotated in MRO (only 
completely annotated complexes are returned)
}
\description{
Assemble a table or MHC protein complexes for a given organism.
}
\examples{
assemble_protein_complex(organism = "mouse")

}
