% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRO_interface.R
\name{assemble_protein_complex}
\alias{assemble_protein_complex}
\title{assemble_protein_complex}
\usage{
assemble_protein_complex(organism_id)
}
\arguments{
\item{organism_id}{Organism for which the lookup should be built (e.g. "human", "mouse", ...)}
}
\value{
a data frame with the MHC complexes annotated in MRO (only completely annotated complexes are returned)
}
\description{
assemble_protein_complex
}
\examples{
org_id <- organism_input("mouse")
assemble_protein_complex(org_id)

}
