% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACUI_interface.R
\name{decode_MAC}
\alias{decode_MAC}
\title{Decode MAC}
\usage{
decode_MAC(MAC)
}
\arguments{
\item{MAC}{multiple allele code (e.g. "A*01:ATJNV")}
}
\value{
decoded MAC
}
\description{
Decode a multiple allele code (MAC) into a list of HLA alleles.
}
\examples{
MAC <- "A*01:ATJNV"
decode_MAC(MAC)

}
