% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MACUI_interface.R
\name{decode_MAC}
\alias{decode_MAC}
\title{decode_MAC}
\usage{
decode_MAC(MAC)
}
\arguments{
\item{MAC}{}
}
\value{
decoded MAC
}
\description{
decode_MAC
}
\examples{
MAC <- "A*01:ATJNV"
decode_MAC(MAC)

}
