% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFND_interface.R
\name{extract_population_id}
\alias{extract_population_id}
\title{extract_population_id}
\usage{
extract_population_id(data)
}
\arguments{
\item{data}{html character string to parse population ids from}
}
\value{
list of population ids present in data
}
\description{
extract_population_id
}
