% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomenclature_queries.R
\name{get_P_group}
\alias{get_P_group}
\title{get_P_group}
\usage{
get_P_group(allele_list)
}
\arguments{
\item{allele_list}{}
}
\value{
Named list of P-groups the input alleles belong to.
}
\description{
get_P_group
}
\examples{
allele_list <- c("DQB1*02:02:01", "DQB1*06:09:01")
get_P_group(allele_list)

}
