% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRO_interface.R
\name{get_valid_organisms}
\alias{get_valid_organisms}
\title{get_valid_organisms}
\usage{
get_valid_organisms()
}
\value{
list of organisms
}
\description{
get the list of organisms that are part of the MRO annotation
}
\examples{
get_valid_organisms()

}
