% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_html.R
\name{getURL}
\alias{getURL}
\title{getURL}
\usage{
getURL(URL, N.TRIES = 1L)
}
\arguments{
\item{URL}{url that will be read}

\item{N.TRIES}{number, how often should the function try to read the html}
}
\value{
returns output of read_html(), a html character string.
}
\description{
getURL
}
