% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRO_interface.R
\docType{data}
\name{human_protein_complex_table}
\alias{human_protein_complex_table}
\title{human_protein_complex_table}
\format{
An object of class \code{data.frame} with 12385 rows and 8 columns.
}
\usage{
human_protein_complex_table
}
\description{
human_protein_complex_table
}
\details{
\code{human_protein_complex_table}: human_protein_complex_table.
}
\examples{
# The human protein complex table is available in the following
# exported variable
human_protein_complex_table
}
\keyword{datasets}
