% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_allele_frequency}
\alias{plot_allele_frequency}
\title{plot_allele_frequency}
\usage{
plot_allele_frequency(allele_frequency)
}
\arguments{
\item{allele_frequency}{}
}
\value{
ggplot
}
\description{
plot_allele_frequency
}
\examples{

# select frequency of given allele
sel_allele_freq <- query_allele_frequencies(hla_selection = "A*02:01", 
hla_sample_size_pattern = "bigger_than", 
hla_sample_size = 10000, standard="g")

plot_allele_frequency(sel_allele_freq)

}
